edit: impl separation
